move virtual_circuit resource & datasource to isolated package
