+++ internal/resources/metal/virtual_circuit/datasource.go
@@ -1,4 +1,4 @@
-package equinix
+package virtual_circuit
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func dataSourceMetalVirtualCircuit() *schema.Resource {
+func DataSource() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceMetalVirtualCircuitRead,
